+++ ChangeLog.md
@@ -8,6 +8,10 @@ Table:
 
 - Fixed issue for querying on identifiers starting with underscore.
 
+Queue:
+
+- Fixed issue that queue service SAS without start time not work.
+
 ## 2023.02 Version 3.22.0
 
 General: